ER-963: Moved API controllers into sub-namespace
