+++ ding_unilogin/src/Controller/Api/ApiController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\ding_unilogin\Controller;
+namespace Drupal\ding_unilogin\Controller\Api;
 
 use Drupal\ding_unilogin\Exception\HttpUnauthorizedException;
